远程连接执行脚本添加了请求头字段
